chore: divide data table to smaller chunks
