+++ packages/ui/src/DataTable/utils/getFieldPathToField.js
@@ -1,0 +1,7 @@
+export const getFieldPathToField = schema => {
+  const fieldPathToField = {};
+  schema.fields.forEach(f => {
+    fieldPathToField[f.path] = f;
+  });
+  return fieldPathToField;
+};
