+++ packages/ui/src/DataTable/utils/index.js
@@ -1,0 +1,51 @@
+import { isEntityClean } from "./isEntityClean";
+import { getSelectedRowsFromEntities } from "./selection";
+import { removeCleanRows } from "./removeCleanRows";
+import getIdOrCodeOrIndex from "./getIdOrCodeOrIndex";
+import computePresets from "./computePresets";
+import { getRecordsFromIdMap } from "./withSelectedEntities";
+import { formatPasteData } from "./formatPasteData";
+import { getFieldPathToField } from "./getFieldPathToField";
+import {
+  defaultParsePaste,
+  getEntityIdToEntity,
+  getFieldPathToIndex,
+  getNumberStrAtEnd,
+  stripNumberAtEnd
+} from "./utils";
+import { getAllRows } from "./getAllRows";
+import { getNewEntToSelect } from "./getNewEntToSelect";
+import { getLastSelectedEntity } from "./getLastSelectedEntity";
+import { getCellInfo } from "./getCellInfo";
+import { getCellCopyText } from "./getCellCopyText";
+import { getRowCopyText } from "./getRowCopyText";
+import { handleCopyHelper } from "./handleCopyHelper";
+import { handleCopyRows } from "./handleCopyRows";
+import { handleCopyColumn } from "./handleCopyColumn";
+import { isBottomRightCornerOfRectangle } from "./isBottomRightCornerOfRectangle";
+
+export {
+  computePresets,
+  defaultParsePaste,
+  formatPasteData,
+  getAllRows,
+  getCellCopyText,
+  getCellInfo,
+  getEntityIdToEntity,
+  getFieldPathToIndex,
+  getFieldPathToField,
+  getIdOrCodeOrIndex,
+  getLastSelectedEntity,
+  getNewEntToSelect,
+  getNumberStrAtEnd,
+  getRecordsFromIdMap,
+  getRowCopyText,
+  getSelectedRowsFromEntities,
+  handleCopyColumn,
+  handleCopyHelper,
+  handleCopyRows,
+  isBottomRightCornerOfRectangle,
+  isEntityClean,
+  removeCleanRows,
+  stripNumberAtEnd
+};
